Remove `build` target from `library` generator
This "fixes" the issue of nx unnecessarily executing the builder for
local library dependencies when their dependant projects are built,
but users who aren't as experienced with nx (or just unsure about the
capabilities of this plugin) may not realize that they can manually add
a `build` target to the workspace.json if they need to.

A `buildable` option should probably be added (mirroring the 1st-party
plugins), with an explanation of the consequences if used unnecessarily.