Ignore Nx config's `targets` key
Nx targets are configured in a project's configuration file like this:

```json
{
  ...
  "targets": {
    "<target-name>": {
      "executor": "<executor-name>",
      ...
    }
    ...
  }
}
```

Previously, this plugin had been using that <target-name> to
determine which type of command to execute. However, users should be
permitted to assign whatever string they'd like to those keys without
affecting the behavior of the invoked executor. This commit fixes
that issue by determining behavior exclusively from the executor
invoked and the options passed to it.